[FIX] set has_image on 'product.product' model [IMP] make has_image stored [REF] improve description and add screenshots [REF] OCA convention
