+++ pos_default_empty_image/models/__init__.py
@@ -1,1 +1,1 @@
-from . import product
+from . import product_product
