+++ pos_default_empty_image/static/src/css/pos_default_empty_image.css
@@ -1,25 +1,8 @@
-/******************************************************************************
-    Point Of Sale - Improve Images module for OpenERP
-    Copyright (C) 2014 GRAP (http://www.grap.coop)
+/*
+    Copyright (C) 2014 - Today: GRAP (http://www.grap.coop)
     @author Julien WESTE
-    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU Affero General Public License as
-    published by the Free Software Foundation, either version 3 of the
-    License, or (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU Affero General Public License for more details.
-
-    You should have received a copy of the GNU Affero General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-******************************************************************************/
-
-/* 
-    New display for 'product-img' and 'product-name' where there isn't image;
+    @author: Sylvain LE GAL (https://twitter.com/legalsylvain)
+    License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 */
 
 .product-img-without-image{
