+++ pos_default_empty_image/static/src/js/widgets.js
@@ -1,7 +1,6 @@
-odoo.define('pos_default_empty_image', function (require) {
+odoo.define('pos_default_empty_image.widgets', function (require) {
     "use strict";
 
-    var models = require('point_of_sale.models');
     var screens = require('point_of_sale.screens');
 
     var core = require('web.core');
@@ -36,16 +35,6 @@ odoo.define('pos_default_empty_image', function (require) {
                 }
                 return cached;
             }
-        },
-    });
-
-    var _super_posmodel = models.PosModel.prototype;
-    models.PosModel = models.PosModel.extend({
-        initialize: function (session, attributes) {
-            var product_model = _.find(this.models, function(model){ return model.model === 'product.product'; });
-            product_model.fields.push('has_image');
-
-            return _super_posmodel.initialize.call(this, session, attributes);
         },
     });
 });
